refactor: move custom scripts logic to its own file
